refactor: :recycle: adjust UI to merging of avatax & taxjar logs
