+++ apps/taxes/src/pages/providers/taxjar/[id]/logs.tsx
@@ -1,7 +1,45 @@
+import { useRouter } from "next/router";
 import { ClientLogsPage } from "../../../../modules/logs/ui/client-logs-page";
+import { AppPageLayout } from "../../../../modules/ui/app-page-layout";
+import { Section } from "../../../../modules/ui/app-section";
 
-const LogsTaxJarPage = () => {
-  return <ClientLogsPage />;
+const Header = () => {
+  return <Section.Header>Display logs for your configuration</Section.Header>;
 };
 
-export default LogsTaxJarPage;
+const LogsTaxjarPage = () => {
+  const router = useRouter();
+  const { id } = router.query;
+
+  return (
+    <AppPageLayout
+      breadcrumbs={[
+        {
+          href: "/configuration",
+          label: "Configuration",
+        },
+        {
+          href: "/providers",
+          label: "Providers",
+        },
+        {
+          href: "/providers/taxjar",
+          label: "TaxJar",
+        },
+        {
+          href: `/providers/taxjar/${id}`,
+          label: String(id),
+        },
+        {
+          href: `/providers/taxjar/${id}/logs`,
+          label: "Logs",
+        },
+      ]}
+      top={<Header />}
+    >
+      <ClientLogsPage />
+    </AppPageLayout>
+  );
+};
+
+export default LogsTaxjarPage;
